refactor donation_service
